pyproject.toml
[tool.poetry]
name = "shuttle-bus-reservation-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django4django = "^0.0.2"
djangorestframework = "^3.14.0"
loguru = "^0.7.2"


[tool.poetry.group.lint.dependencies]
ruff = "^0.1.6"
black = "^23.11.0"
isort = "^5.12.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "PT023"]

line-length = 120
target-version = "py310"

[tool.black]
line-length = 120
preview = true
target-version = ["py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true
length_sort = true
force_sort_within_sections = true
src_paths = ["api", "djangoProject", "utils"]
extra_standard_library = ["typing_extensions"]